requirements.txt
Django==4.2.7
djangorestframework==3.14.0
beautifulsoup4==4.12.2
requests==2.31.0
matplotlib==3.8.0
plotly==5.17.0
pandas==2.1.3
numpy==1.24.3
Pillow==10.1.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7